Make the NextLevelOverlay non-draggable to make clicking "continue" easier
After demoing NumptyPhysics, I noticed that users had a hard time clicking
the "continue" button on the NextLevelOverlay, because they started dragging
the window, which makes it difficult to hit the button. I've added a new
variable to the Overlay class (dragging_allowed) that disables dragging if
we do not want this. It still defaults to "true" for IconOverlay, but we
set it to "false" for the NextLevelOverlay.

This way, the "continue" button is easier to hit. And moving around the
NextLevelOverlay is not needed for anything, so I think this is okay :)